Fix: convert char max length to Long

+++ src/main/java/ma/ilias/dbmanagementbe/metadata/dto/common/IColumnDataTypeDefinition.java
@@ -3,7 +3,7 @@ package ma.ilias.dbmanagementbe.metadata.dto.common;
 public interface IColumnDataTypeDefinition {
     String getDataType();
 
-    Integer getCharacterMaxLength();
+    Long getCharacterMaxLength();
 
     Integer getNumericPrecision();
